EduPilot/requirements.txt
openenv-core[core]
fastapi
uvicorn
isort
black
httpx
jsonschema
python-dotenv
langchain_core
langchain_groq